--- src/jobs/jobs.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { JobsController } from './jobs.controller';
-
-describe('Jobs Controller', () => {
-  let controller: JobsController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [JobsController],
-    }).compile();
-
-    controller = module.get<JobsController>(JobsController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
